Early/alpha release of PFCustodian v0.1.
PFCustodian is a set of scripts to make use of the Pathfinder Character Sheet
more convenient on Roll20.  It's stated goal is to automate as much of the
numeric book-keeping as possible in order to both ensure accuracy and lower
workload on the players.

It is explicitly and purposefully tied to both the Pathfinder game system
and to the default Pathfinder Character Sheet on Roll20.